[#79] refactor: agora 조회 로직 수정

+++ src/main/java/com/attica/athens/domain/agora/dto/SimpleParticipants.java
@@ -1,8 +1,8 @@
 package com.attica.athens.domain.agora.dto;
 
 public record SimpleParticipants(
-        int pros,
-        int cons,
-        int observer
+        Long pros,
+        Long cons,
+        Long observer
 ) {
 }
